// Files/src/js/models/enums/game-type.ts
export enum GameType {
    UNKNOWN = 0,
    VS_AI = 1,
    VS_FRIEND = 2,
    TUTORIAL = 4,
    ARENA = 5,
    TEST = 6,
    RANKED = 7,
    CASUAL = 8,
    TAVERNBRAWL = 16,
    TB_1P_VS_AI = 17,
    TB_2P_COOP = 18,
    FSG_BRAWL_VS_FRIEND = 19,
    FSG_BRAWL = 20,
    FSG_BRAWL_1P_VS_AI = 21,
    FSG_BRAWL_2P_COOP = 22,
    TOURNAMENT = 23,
}